Quorum: Fix printing of size_t.  Not all platforms have size_t == int.
Mercurial revision: a63a791287f0eb8331f58c2b4b1458e34e12ce25

+++ membership/quorumd/quorumd.c
@@ -422,7 +422,7 @@ on_listen(GIOChannel *ch, GIOCondition condition, gpointer data)
 			close(csock);
 			return TRUE;
 		}
-		quorum_debug(LOG_DEBUG, "version:%s(%d)",buf,len);
+		quorum_debug(LOG_DEBUG, "version:%s(%d)",buf,(int)len);
 		fns = get_protocol(buf);
 		if(fns != NULL) {
 			fns->on_connect(csock,session,CN);
